rename multiFidelityFunction.y -> multi_fidelity_function.py
