+++ mf2/__init__.py
@@ -7,7 +7,7 @@ A collection of analytical functions with 2 or more available fidelities.
 """
 import pkg_resources
 
-from .multiFidelityFunction import MultiFidelityFunction, invert
+from .multi_fidelity_function import MultiFidelityFunction, invert
 from .borehole import borehole
 from .currin import currin
 from .park91a import park91a
