+++ mf2/bohachevsky.py
@@ -9,7 +9,7 @@
 
 import numpy as np
 
-from .multiFidelityFunction import MultiFidelityFunction
+from .multi_fidelity_function import MultiFidelityFunction
 
 
 def bohachevsky_hf(xx):
