+++ mf2/borehole.py
@@ -29,7 +29,7 @@ General Public License for more details.
 
 import numpy as np
 
-from .multiFidelityFunction import MultiFidelityFunction
+from .multi_fidelity_function import MultiFidelityFunction
 
 
 _tau = 2*np.pi
